Address token.place metrics review feedback
